improved proof serialization/deserialization
