+++ verifier/src/channel.rs
@@ -4,36 +4,37 @@
 // LICENSE file in the root directory of this source tree.
 
 use common::{
-    errors::VerifierError,
-    proof::{OodEvaluationFrame, Queries, StarkProof},
-    utils, Air, ComputationContext, EvaluationFrame, ProofOptions, PublicCoin,
+    errors::VerifierError, proof::StarkProof, utils, Air, ComputationContext, EvaluationFrame,
+    ProofOptions, PublicCoin,
 };
 use crypto::{BatchMerkleProof, Hasher, MerkleTree};
 use fri::{PublicCoin as FriPublicCoin, VerifierChannel as FriVerifierChannel};
-use math::{
-    field::{FieldElement, StarkField},
-    utils::read_elements_into_vec,
-};
+use math::field::{FieldElement, StarkField};
 use std::convert::TryInto;
-use std::marker::PhantomData;
 
 // TYPES AND INTERFACES
 // ================================================================================================
 
 pub struct VerifierChannel<B: StarkField, E: FieldElement + From<B>, H: Hasher> {
     context: ComputationContext,
+    // trace queries
     trace_root: H::Digest,
-    trace_queries: Queries,
+    trace_proof: BatchMerkleProof<H>,
+    trace_states: Vec<Vec<B>>,
+    // constraint queries
     constraint_root: H::Digest,
-    constraint_queries: Queries,
-    ood_frame: OodEvaluationFrame,
+    constraint_proof: BatchMerkleProof<H>,
+    constraint_evaluations: Vec<Vec<E>>,
+    // FRI proof
     fri_roots: Vec<H::Digest>,
     fri_layer_proofs: Vec<BatchMerkleProof<H>>,
     fri_layer_queries: Vec<Vec<E>>,
     fri_remainder: Vec<E>,
     fri_partitioned: bool,
+    // query seed
     query_seed: H::Digest,
-    _base_element: PhantomData<B>,
+    // out-of-domain evaluation frame
+    ood_frame: EvaluationFrame<E>,
 }
 
 // VERIFIER CHANNEL IMPLEMENTATION
@@ -58,6 +59,30 @@ where
             .parse::<H>(num_fri_layers)
             .map_err(|err| VerifierError::ProofDeserializationError(err.to_string()))?;
 
+        // --- parse trace queries ----------------------------------------------------------------
+        let (trace_proof, trace_states) = proof
+            .trace_queries
+            .parse::<H, B>(air.lde_domain_size(), air.trace_width())
+            .map_err(|err| {
+                VerifierError::ProofDeserializationError(format!(
+                    "trace query deserialization failed: {}",
+                    err.to_string()
+                ))
+            })?;
+
+        // --- parse constraint evaluation queries ------------------------------------------------
+        let evaluations_per_leaf = utils::evaluations_per_leaf::<E, H>();
+        let num_leaves = air.lde_domain_size() / evaluations_per_leaf;
+        let (constraint_proof, constraint_evaluations) = proof
+            .constraint_queries
+            .parse::<H, E>(num_leaves, evaluations_per_leaf)
+            .map_err(|err| {
+                VerifierError::ProofDeserializationError(format!(
+                    "constraint evaluation query deserialization failed: {}",
+                    err.to_string()
+                ))
+            })?;
+
         // --- parse FRI proofs -------------------------------------------------------------------
         let fri_partitioned = proof.fri_proof.is_partitioned();
         let fri_remainder = proof
@@ -74,66 +99,50 @@ where
         let query_seed =
             build_query_seed::<H>(&fri_roots, proof.pow_nonce, &air.context().options())?;
 
+        // --- parse out-of-domain evaluation frame -----------------------------------------------
+        let ood_frame = proof
+            .ood_frame
+            .parse(air.trace_width())
+            .map_err(|err| VerifierError::ProofDeserializationError(err.to_string()))?;
+
         Ok(VerifierChannel {
             context: air.context().clone(),
-            ood_frame: proof.ood_frame,
+            // trace queries
             trace_root,
-            trace_queries: proof.trace_queries,
+            trace_proof,
+            trace_states,
+            // constraint queries
             constraint_root,
-            constraint_queries: proof.constraint_queries,
+            constraint_proof,
+            constraint_evaluations,
+            // FRI proof
             fri_roots,
             fri_layer_proofs,
             fri_layer_queries,
             fri_remainder,
             fri_partitioned,
+            // query seed
             query_seed,
-            _base_element: PhantomData,
+            // out-of-domain evaluation frame
+            ood_frame,
         })
     }
 
-    /// Returns trace polynomial evaluations at OOD points z and z * g, where g is the generator
-    /// of the LDE domain.
-    pub fn read_ood_frame(&self) -> Result<EvaluationFrame<E>, VerifierError> {
-        let current = match read_elements_into_vec(&self.ood_frame.trace_at_z1) {
-            Ok(elements) => {
-                if elements.len() != self.context.trace_width() {
-                    return Err(VerifierError::OodFrameDeserializationFailed);
-                }
-                elements
-            }
-            Err(_) => return Err(VerifierError::OodFrameDeserializationFailed),
-        };
-        let next = match read_elements_into_vec(&self.ood_frame.trace_at_z2) {
-            Ok(elements) => {
-                if elements.len() != self.context.trace_width() {
-                    return Err(VerifierError::OodFrameDeserializationFailed);
-                }
-                elements
-            }
-            Err(_) => return Err(VerifierError::OodFrameDeserializationFailed),
-        };
-
-        Ok(EvaluationFrame { current, next })
+    /// Returns trace polynomial evaluations at out-of-domain points z and z * g, where
+    /// g is the generator of the LDE domain.
+    pub fn read_ood_frame(&self) -> &EvaluationFrame<E> {
+        &self.ood_frame
     }
 
     /// Returns trace states at the specified positions. This also checks if the
     /// trace states are valid against the trace commitment sent by the prover.
-    pub fn read_trace_states(&self, positions: &[usize]) -> Result<Vec<Vec<B>>, VerifierError> {
-        // deserialize query bytes into a set of trace states at the specified positions
-        // and corresponding Merkle paths
-        // TODO: avoid cloning
-        let (merkle_paths, trace_states) = self
-            .trace_queries
-            .clone()
-            .parse::<H, B>(self.context.lde_domain_size(), self.context.trace_width())
-            .map_err(|_err| VerifierError::TraceQueryDeserializationFailed)?;
-
+    pub fn read_trace_states(&self, positions: &[usize]) -> Result<&[Vec<B>], VerifierError> {
         // make sure the states included in the proof correspond to the trace commitment
-        if !MerkleTree::verify_batch(&self.trace_root, positions, &merkle_paths) {
+        if !MerkleTree::verify_batch(&self.trace_root, positions, &self.trace_proof) {
             return Err(VerifierError::TraceQueryDoesNotMatchCommitment);
         }
 
-        Ok(trace_states)
+        Ok(&self.trace_states)
     }
 
     /// Returns constraint evaluations at the specified positions. This also checks if the
@@ -143,30 +152,20 @@ where
         positions: &[usize],
     ) -> Result<Vec<E>, VerifierError> {
         let evaluations_per_leaf = utils::evaluations_per_leaf::<E, H>();
-        let num_leaves = self.context.lde_domain_size() / evaluations_per_leaf;
-        // deserialize query bytes into a set of constraint evaluations at the specified positions
-        // and corresponding Merkle paths
-        // TODO: avoid cloning
-        let (merkle_paths, constraint_evaluations) = self
-            .constraint_queries
-            .clone()
-            .parse::<H, E>(num_leaves, evaluations_per_leaf)
-            .map_err(|_err| VerifierError::ConstraintQueryDeserializationFailed)?;
-
         let c_positions = utils::map_trace_to_constraint_positions(positions, evaluations_per_leaf);
-        if !MerkleTree::verify_batch(&self.constraint_root, &c_positions, &merkle_paths) {
+        if !MerkleTree::verify_batch(&self.constraint_root, &c_positions, &self.constraint_proof) {
             return Err(VerifierError::ConstraintQueryDoesNotMatchCommitment);
         }
 
         // build constraint evaluation values from the leaves of constraint Merkle proof
         let mut evaluations: Vec<E> = Vec::with_capacity(positions.len());
         for &position in positions.iter() {
-            // TODO: position computation should be in common
             let leaf_idx = c_positions
                 .iter()
                 .position(|&v| v == position / evaluations_per_leaf)
                 .unwrap();
-            evaluations.push(constraint_evaluations[leaf_idx][position % evaluations_per_leaf]);
+            evaluations
+                .push(self.constraint_evaluations[leaf_idx][position % evaluations_per_leaf]);
         }
 
         Ok(evaluations)
